refactor(query): Extract join methods to planner/joins.rs
Move 5 join methods (~2,231 lines) from builder.rs to joins.rs:
apply_joins, build_scan_mapping_for_join, apply_filter_relation_join,
apply_multi_level_sub_joins, apply_multi_level_filter_joins.

builder.rs: 4,398 → 2,146 lines
